Always use generic implementations of the ApiProxyImpl and RequestManager
